# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API документация",
        "version": "1.0.0"
    },
    "paths": {
        "/api/rent/car": {
            "get": {
                "summary": "Регистрация аренды автомобиля пользователем",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "query",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "5"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/terminate/rentcar": {
            "get": {
                "summary": "Прекращение аренды автомобиля",
                "parameters": [
                    {
                        "name": "car_id",
                        "in": "query",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "5"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}